Various fixes and improvements for the voting tool
