+++ .github/workflows/closeVote.yml
@@ -51,7 +51,6 @@ jobs:
           # Loading the default branch so we use the last version of the mailmap
           # rather than getting stuck to when the vote PR was open.
           ref: ${{ github.event.repository.default_branch }}
-          persist-credentials: true  # we need the credentials to push the new vote branch
       - name: Download nodejs/node mailmap file
         run:
           curl -L https://raw.githubusercontent.com/nodejs/node/main/.mailmap >>
@@ -69,20 +68,46 @@ jobs:
       - name: Attempt closing the vote
         id: vote-summary
         run: |
-          EOF=$(dd if=/dev/urandom bs=15 count=1 status=none | base64)
-          echo "markdown<<$EOF" >> "$GITHUB_OUTPUT"
-          ./votes/initiateNewVote/decryptPrivateKeyAndCloseVote.mjs \
-          --remote origin --branch "${{ steps.branch.outputs.head }}" \
-          --fromCommit "FETCH_HEAD~${{ steps.nb-of-commits.outputs.minusOne }}" \
-          --toCommit "FETCH_HEAD" \
-          --prURL "${{ steps.pr-url.outputs.URL }}" \
-          --save-markdown-summary summaryComment.md \
-          --comments "$COMMENTS" --commit-json-summary >> "$GITHUB_OUTPUT"
-          echo "$EOF" >> "$GITHUB_OUTPUT"
+          {
+            EOF=$(dd if=/dev/urandom bs=15 count=1 status=none | base64)
+            echo "markdown<<$EOF"
+            ./votes/initiateNewVote/decryptPrivateKeyAndCloseVote.mjs \
+              --remote origin --branch "${{ steps.branch.outputs.head }}" \
+              --fromCommit "FETCH_HEAD~${{ steps.nb-of-commits.outputs.minusOne }}" \
+              --toCommit "FETCH_HEAD" \
+              --prURL "${{ steps.pr-url.outputs.URL }}" \
+              --save-markdown-summary summaryComment.md \
+              --comments "$COMMENTS" --commit-json-summary
+            echo "$EOF"
+          } >> "$GITHUB_OUTPUT"
         env:
           COMMENTS: ${{ steps.comments.outputs.comments }}
+      - name: Install ghcommit
+        run: go install github.com/planetscale/ghcommit@8c6d9af75a7814768ce871cde246224d45bd8c04
       - name: Push to the PR branch
-        run: git push origin "HEAD:${{ steps.branch.outputs.head }}"
+        run: |
+          GH_COMMIT_PATH="$(go env GOPATH)/bin/ghcommit" COMMIT_MESSAGE="$(
+            git log -1 HEAD --pretty=format:%B
+          )" SHA="$(
+            git rev-parse HEAD^
+          )" DELETED_FILES="$(
+            git show HEAD --name-only --diff-filter=D --pretty=format:
+          )" ADDED_FILES="$(
+            git show HEAD --name-only --diff-filter=d --pretty=format:
+          )" node --input-type=module <<'EOF'
+            import { spawnSync } from "node:child_process";
+            const {GH_COMMIT_PATH, COMMIT_MESSAGE, SHA, DELETED_FILES, ADDED_FILES} = process.env;
+            spawnSync(GH_COMMIT_PATH, [
+              '-r', ${{ toJSON(github.repository) }},
+              '-b', ${{ toJSON(steps.branch.outputs.head) }},
+              '-m', COMMIT_MESSAGE,
+              '--sha', SHA,
+              ...DELETED_FILES.split('\n').filter(Boolean).flatMap(file => ['--delete', file]),
+              ...ADDED_FILES.split('\n').filter(Boolean).flatMap(file => ['--add', file]),
+            ], { stdio: 'inherit' });
+          EOF
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
       - name: Publish vote summary comment
         run: |
           gh pr comment "${{ steps.pr-url.outputs.URL }}" --body-file summaryComment.md
